people mastodon_url: use rel="me"
from "link verification" at https://docs.joinmastodon.org/user/profile/

> If you put a link in your profile metadata, Mastodon checks if the
linked page links back to your Mastodon profile. If so, you get a
verification checkmark next to that link, since you are confirmed as the
owner.

> Behind the scenes, Mastodon checks for the rel="me" attribute on the
link back. Likewise, Mastodon puts rel="me" on the links within profile
metadata.

This scenario is slightly different since the person is not *the* owner
of the torproject people page, but I think from context it's clear that
this is verifying that the mastodon profile belongs to *someone* on the
torproject people page.