Small corrections
the sponsor model does exist, and apparently contains the same thing as
the supporter model. however, all of the files in the `supporters`
directory use the `supporter` model, so we should stick with the same.

I tried opening up the URL with www and it failed to connect. Their
website works well without www so let's use that instead.